QR/requirements.txt
Flask==2.1.2
qrcode
pillow
